add class diagram and some methods

+++ src/com/hospital/models/TimeSlot.java
@@ -2,27 +2,27 @@ package com.hospital.models;
 
 public class TimeSlot {
 
-	private int startTime;
-	private int endTime;
+	private double startTime;
+	private double endTime;
 	
-	public TimeSlot(int startTime,int endTime) {
+	public TimeSlot(double startTime,double endTime) {
 		this.startTime = startTime;
 		this.endTime = endTime;
 	}
 
-	public int getStartTime() {
+	public double getStartTime() {
 		return startTime;
 	}
 
-	public void setStartTime(int startTime) {
+	public void setStartTime(double startTime) {
 		this.startTime = startTime;
 	}
 
-	public int getEndTime() {
+	public double getEndTime() {
 		return endTime;
 	}
 
-	public void setEndTime(int endTime) {
+	public void setEndTime(double endTime) {
 		this.endTime = endTime;
 	}
